feat: 프로젝트 이력 status 필드 추가
